+++ app/src/main/java/edu/hanover/hc24_luuk_crawford_senior_project/data/MenuItem.kt
@@ -7,7 +7,8 @@ package edu.hanover.hc24_luuk_crawford_senior_project.data
  * @param customizationType category of customization
  * @param imageLink link to image (at least jpeg works)
  */
-data class MenuItem(val name: String = "default item", val id: Int = 0,
-                    val customizationType: String = "default type", val imageLink: String = "https://i.imgur.com/BDa36Zp.jpeg")
+data class MenuItem(
+    val name: String = "default item", val id: Long = 0L,
+    val customizationType: String = "default type", val imageLink: String = "https://i.imgur.com/BDa36Zp.jpeg")
