+++ app/src/main/java/edu/hanover/hc24_luuk_crawford_senior_project/data/UserOrder.kt
@@ -9,7 +9,7 @@ data class UserOrder(
     var user: String,
     var userID: Int,
     val orderID: Int,
-    var itemID: Int,
+    var itemID: Long,
     var customization: Customization,
     var itemStatus: ItemStatus,
     var orderTime: Date?,
